Changed int to double because it's based on division and there's no reason to not be a little more accurate right now. This is just one part of this second attempt to support multiple platforms for development. Worked! Yay!

+++ include/managers/SceneManager.h
@@ -12,13 +12,13 @@ namespace TIE
 	{
 		public:
 			void render();
-			int getFPS();
+			double getFPS();
 
 			~SceneManager();
 			SceneManager();
 		private:
 			sf::Clock& clock;
-			int fps;
+			double fps;
 			GlobalId sceneViewId;
 
 			SceneManager(const SceneManager&);
